refactor: keep in-process host client internal instead of public SDK
Move the app-server embedding facade from the public sdk/ package to
internal/sdk and drop the external examples. The wuu exec and wuu app-server
entrypoints still share the consolidated in-process client, but the project no
longer commits to a public Go embedding API until a concrete external consumer
exists.
