+++ internal/sdk/session.go
